Initial commit. Most configs based on theprimeagen
I am exploring neo vim. The description primeagen gave was that the
limiting factor when coding is being able to get the commands into the
machine. Most of my lag is not that but it would be nice to not have to
move my hands off the keyboard so much.